Added command line option for specifying covariate SNP, and genotype loading code
No functional changes yet, just loading code. Verified that results
are the same for the tested marker.

Also added code to output coefficients and p-values for all
fixed effect terms in the model, including intercept. I can't be
certain conceptually that I am doing the math right, but I am
satisfied with the output which shows the intercept is highly
significant with p-value = 0 for all tests, and first two PC
covariates are not significant pretty much ever. The 3rd PC
however is significant for a handful of tests. This is consistent
with my expectations as we have found that including 1 or 2 PCs
does not make any meaningful changes in the outcome of the
analysis, but 3 PCs does make some differences. It was never
clear however that these differences warranted including 3 PCs
in the model. We never had a way of knowing the 1st two PCs had
no effect at all, but if I'm doing this right, the p-values
clearly show they do not. This is consistent with the kinship
matrix completely describing the major population structure. The
3rd PC may be picking up on outlying individuals, only a 3
dimensional PC plot would make this clear (or not) I think. This
suggests however that the 3rd PC alone perhaps should be included
in the model. It may be the case though that all it is doing is
decreasing power by picking up on individuals more diverged from
the main population that also have quite different phenotypes as
a result of this. All this however is outside the present scope
of trying to add an additional SNP as a covariate and then
SNP-SNP interaction terms.