x1 and x1t vectors turned to matrices
I now have the x1 and x1t vectors converted to matrices and the
call which does the computation of x1t = t(eLvecs) %*% x1 now
performing the operation for matrices rather than just a single
marker vector.

Nothing done so far should change any of the results since the
subsequent calls for solving the model are not using the additional
columns of x1 and x1t. I have verified the same results are still
produced at this stage. Most likely any next change will change
the outcome, so committing this far now.