Successfully modified to include additional genetic terms in the model
Current only adds a second specific marker that is specified on the
command line. No GxG terms are added at this time but now that this
is working, doing so should be relatively straightforward. The hard
part (I hope) was expanding the single marker genetic term, expressed
as a vector, into a matrix, and modifying the code that fills in
the XDX matrix and XDy vector with the values that are dependent
on these. I believe now that has been successfully done.