PVE of genetic model
Attempt to calculate total PVE of the fixed-effect genetic
component of the model by first
fitting the model with all terms but the genotypes set to random
values, and computing the residual variance as per how I've been
doing that I thought was working. Then for calculating PVE(genetic),
I just take PVE(model) and subtract the PVE found for the model
when the genotypes are random. I do the random thing a 1000 times
and average over the trials to get a stable estimate.

This seems reasonable, if a bit hackish, but has the advantage in
that the model with effectively zero genetic effects has the same
degrees of freedom. However, the estimate for what the kinship
matrix explains alone seems low, the PVE(genetic) calculated is
much higher than the sum of the PVEs across genetic terms for
models with two fixed SNP covariates (the two major peaks) but no
interaction terms. (haven't tried with interaction terms yet)

Either (a) I'm getting the residual variance wrong (probably), or
(b) the most significant SNPs included in the model end up enabling
the kinship random effects term to explain a great deal more
variance than it does with random genotypes. The main advance of
EMMAX over EMMA is it's performance, which it achieves by only
fitting the random effects terms without fixed genetic effects and
reuses this for every marker tested in the genome-wide scan rather
than re-estimate what it would be with the additional genetic
fixed effect terms. This is reasonable under the assumption that
no single marker explains a huge amount of variance. Unfortunately,
in our case, one of them does. This raises questions about how to
proceed.