Change to deal with case that a covariate SNP is not found in input
I just substitute random values for the SNP's genotype. The
covariate then effectively describes nothing meaningful in the
phenotype but still appears in the output making it regularized
with respect to other population subgroupings that do contain the
SNP. Also, this avoids a lot of changes to the code to deal with
this case otherwise.